Add support for LG P970
This should make porting newer version of CyanogenMod / Android to old unsupported devices much easier.
We define the product, for example

"PRODUCT_LG_P970 := true"

and

"COMMON_GLOBAL_CFLAGS += -DPRODUCT_LG_P970"

in the boardconfig

Here we check, if it is defined.
if its defined, we use the known code, with which the build works.

if its not defined, we use the default code

Change-Id: I39e633fbed12b6fe6dc23d5f061f0c504214378b